refactor(rust): extract ai provider routing behind a trait
Provider routing was a match expression duplicated across three commands
(ai_generate, ai_test_provider_key, ai_list_provider_models). Replace it
with an AiProvider trait + a single provider_for() registry, mirroring the
existing ApplierRegistry pattern.

- New commands/ai_provider.rs: trait AiProvider { stream_chat, test_key,
  list_models } with Ollama/OpenAi/Anthropic/Gemini impls and the four
  stream_* functions
- commands/ai.rs commands are now thin dispatchers; get_provider_key is
  pub(crate) and each provider resolves its own key

Adding a provider is now one struct + one impl + one match arm, with no
edits to the command layer.
