+++ apps/tauri/src-tauri/src/commands/ai.rs
@@ -5,6 +5,8 @@ use tauri::{AppHandle, Emitter, Manager};
 use crate::credentials::CredentialStore;
 use crate::jobs::JobTracker;
 
+use super::ai_provider::{provider_for, AiGenerateRequest};
+
 fn uuid_v4() -> String {
     use std::time::{SystemTime, UNIX_EPOCH};
     let t = SystemTime::now()
@@ -16,36 +18,15 @@ fn uuid_v4() -> String {
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct AiMessage {
-    pub role: String,
-    pub content: String,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-#[allow(dead_code)]
-pub struct AiGenerateRequest {
-    pub model: String,
-    pub messages: Vec<AiMessage>,
-    pub locale: Option<String>,
-    pub temperature: Option<f64>,
-    pub max_tokens: Option<u64>,
-    pub stream: Option<bool>,
-    pub provider: Option<String>,
-    pub base_url: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
 pub struct AiEmbedRequest {
     pub text: String,
     pub model: Option<String>,
 }
 
-/// Stream an AI generation from Ollama.
+/// Stream an AI generation from the configured provider.
 #[tauri::command]
 pub async fn ai_generate(app: AppHandle, req: AiGenerateRequest) -> Value {
-    let provider = req.provider.as_deref().unwrap_or("ollama").to_string();
+    let provider_name = req.provider.as_deref().unwrap_or("ollama").to_string();
 
     let job_id = uuid_v4();
     app.state::<Mutex<JobTracker>>()
@@ -56,26 +37,8 @@ pub async fn ai_generate(app: AppHandle, req: AiGenerateRequest) -> Value {
     let app_clone = app.clone();
 
     tauri::async_runtime::spawn(async move {
-        let result = match provider.as_str() {
-            "openai" | "openai-compatible" => {
-                let api_key = get_provider_key(&app_clone, &provider).unwrap_or_default();
-                let base_url = req.base_url.as_deref().unwrap_or("https://api.openai.com/v1").to_string();
-                stream_openai_chat(&app_clone, &base_url, &api_key, &job_id_clone, &req).await
-            }
-            "anthropic" => {
-                let api_key = get_provider_key(&app_clone, "anthropic").unwrap_or_default();
-                stream_anthropic_chat(&app_clone, &api_key, &job_id_clone, &req).await
-            }
-            "gemini" => {
-                let api_key = get_provider_key(&app_clone, "gemini").unwrap_or_default();
-                stream_gemini_chat(&app_clone, &api_key, &job_id_clone, &req).await
-            }
-            _ => {
-                let base = std::env::var("OLLAMA_HOST")
-                    .unwrap_or_else(|_| "http://127.0.0.1:11434".to_string());
-                stream_ollama_chat(&app_clone, &base, &job_id_clone, &req).await
-            }
-        };
+        let provider = provider_for(&provider_name);
+        let result = provider.stream_chat(&app_clone, &job_id_clone, &req).await;
 
         if let Err(e) = result {
             let _ = app_clone.emit(
@@ -92,112 +55,7 @@ pub async fn ai_generate(app: AppHandle, req: AiGenerateRequest) -> Value {
     json!({ "jobId": job_id })
 }
 
-async fn stream_ollama_chat(
-    app: &AppHandle,
-    base: &str,
-    job_id: &str,
-    req: &AiGenerateRequest,
-) -> Result<(), String> {
-    let messages = serde_json::to_value(&req.messages.iter().map(|m| json!({ "role": m.role, "content": m.content })).collect::<Vec<_>>()).unwrap_or(json!([]));
-
-    let mut body = json!({
-        "model": req.model,
-        "messages": messages,
-        "stream": true,
-    });
-    let mut options = serde_json::Map::new();
-    if let Some(t) = req.temperature {
-        options.insert("temperature".to_string(), json!(t));
-    }
-    if let Some(mt) = req.max_tokens {
-        options.insert("num_predict".to_string(), json!(mt));
-    }
-    if !options.is_empty() {
-        body["options"] = Value::Object(options);
-    }
-
-    let mut response = reqwest::Client::builder()
-        .timeout(std::time::Duration::from_secs(300))
-        .build()
-        .map_err(|e| e.to_string())?
-        .post(format!("{base}/api/chat"))
-        .json(&body)
-        .send()
-        .await
-        .map_err(|e| format!("Ollama unreachable: {e}"))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body_text = response.text().await.unwrap_or_default();
-        return Err(format!("Ollama {status}: {body_text}"));
-    }
-
-    let mut line_buf = String::new();
-
-    loop {
-        // Check if job was cancelled before reading next chunk
-        if let Some(job) = app.state::<Mutex<JobTracker>>().lock().get(job_id) {
-            if job.status == crate::jobs::JobStatus::Cancelled {
-                // Force close the connection aggressively
-                let _ = response.error_for_status_ref();
-                return Err("Job cancelled".to_string());
-            }
-        }
-
-        match response.chunk().await {
-            Ok(Some(bytes)) => {
-                line_buf.push_str(&String::from_utf8_lossy(&bytes));
-
-                while let Some(nl) = line_buf.find('\n') {
-                    let line = line_buf[..nl].trim().to_string();
-                    line_buf = line_buf[nl + 1..].to_string();
-
-                    if line.is_empty() {
-                        continue;
-                    }
-
-                    let event: Value = match serde_json::from_str(&line) {
-                        Ok(v) => v,
-                        Err(_) => continue,
-                    };
-
-                    let delta = event
-                        .get("message")
-                        .and_then(|m| m.get("content"))
-                        .and_then(|c| c.as_str())
-                        .unwrap_or("");
-                    let done = event
-                        .get("done")
-                        .and_then(|d| d.as_bool())
-                        .unwrap_or(false);
-
-                    let _ = app.emit(
-                        "ai:stream",
-                        json!({ "jobId": job_id, "delta": delta, "done": done }),
-                    );
-
-                    if done {
-                        app.state::<Mutex<JobTracker>>()
-                            .lock()
-                            .complete(job_id, json!({ "done": true }));
-                        return Ok(());
-                    }
-                }
-            }
-            Ok(None) => break,
-            Err(e) => return Err(format!("Stream error: {e}")),
-        }
-    }
-
-    let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": "", "done": true }));
-    app.state::<Mutex<JobTracker>>()
-        .lock()
-        .complete(job_id, json!({ "done": true }));
-
-    Ok(())
-}
-
-fn get_provider_key(app: &AppHandle, provider: &str) -> Option<String> {
+pub(crate) fn get_provider_key(app: &AppHandle, provider: &str) -> Option<String> {
     let store = app.state::<Mutex<CredentialStore>>();
     let guard = store.lock();
     guard
@@ -245,72 +103,10 @@ pub async fn ai_test_provider_key(app: AppHandle, provider: String) -> Value {
         Err(e) => return json!({ "success": false, "error": format!("Failed to create client: {}", e) }),
     };
 
-    let result = match provider.as_str() {
-        "openai" | "openai-compatible" => {
-            let resp = client
-                .get("https://api.openai.com/v1/models")
-                .bearer_auth(&api_key)
-                .send()
-                .await;
-            match resp {
-                Ok(r) => {
-                    if r.status().is_success() {
-                        json!({ "success": true })
-                    } else {
-                        let status = r.status();
-                        json!({ "success": false, "error": format!("API returned status: {}", status) })
-                    }
-                }
-                Err(e) => json!({ "success": false, "error": format!("Request failed: {}", e) }),
-            }
-        }
-        "anthropic" => {
-            let resp = client
-                .post("https://api.anthropic.com/v1/messages")
-                .header("x-api-key", &api_key)
-                .header("anthropic-version", "2023-06-01")
-                .header("content-type", "application/json")
-                .json(&serde_json::json!({
-                    "model": "claude-3-haiku-20240307",
-                    "max_tokens": 1,
-                    "messages": [{"role": "user", "content": "test"}]
-                }))
-                .send()
-                .await;
-            match resp {
-                Ok(r) => {
-                    if r.status().is_success() || r.status() == 400 {
-                        // 400 is OK - it means key is valid but request was malformed (we sent minimal data)
-                        json!({ "success": true })
-                    } else {
-                        let status = r.status();
-                        json!({ "success": false, "error": format!("API returned status: {}", status) })
-                    }
-                }
-                Err(e) => json!({ "success": false, "error": format!("Request failed: {}", e) }),
-            }
-        }
-        "gemini" => {
-            let resp = client
-                .get(format!("https://generativelanguage.googleapis.com/v1/models?key={}", api_key))
-                .send()
-                .await;
-            match resp {
-                Ok(r) => {
-                    if r.status().is_success() {
-                        json!({ "success": true })
-                    } else {
-                        let status = r.status();
-                        json!({ "success": false, "error": format!("API returned status: {}", status) })
-                    }
-                }
-                Err(e) => json!({ "success": false, "error": format!("Request failed: {}", e) }),
-            }
-        }
-        _ => json!({ "success": false, "error": "Unknown provider" }),
-    };
-
-    result
+    match provider_for(&provider).test_key(&client, &api_key).await {
+        Ok(()) => json!({ "success": true }),
+        Err(e) => json!({ "success": false, "error": e }),
+    }
 }
 
 #[tauri::command]
@@ -328,428 +124,7 @@ pub async fn ai_list_provider_models(app: AppHandle, provider: String) -> Value 
         Err(_) => return json!([]),
     };
 
-    match provider.as_str() {
-        "openai" | "openai-compatible" => {
-            let resp = client
-                .get("https://api.openai.com/v1/models")
-                .bearer_auth(&api_key)
-                .send()
-                .await;
-            if let Ok(r) = resp {
-                if let Ok(body) = r.json::<serde_json::Value>().await {
-                    if let Some(data) = body.get("data").and_then(|d| d.as_array()) {
-                        let models: Vec<Value> = data
-                            .iter()
-                            .filter_map(|m| m.get("id").and_then(|id| id.as_str()))
-                            .filter(|id| {
-                                id.starts_with("gpt-")
-                                    || id.starts_with("o1")
-                                    || id.starts_with("o3")
-                                    || id.starts_with("o4")
-                            })
-                            .map(|id| json!({ "name": id }))
-                            .collect();
-                        return json!(models);
-                    }
-                }
-            }
-            json!([])
-        }
-        "anthropic" => {
-            let resp = client
-                .get("https://api.anthropic.com/v1/models")
-                .header("x-api-key", &api_key)
-                .header("anthropic-version", "2023-06-01")
-                .send()
-                .await;
-            if let Ok(r) = resp {
-                if let Ok(body) = r.json::<serde_json::Value>().await {
-                    if let Some(data) = body.get("data").and_then(|d| d.as_array()) {
-                        let models: Vec<Value> = data
-                            .iter()
-                            .filter_map(|m| m.get("id").and_then(|id| id.as_str()))
-                            .filter(|id| id.starts_with("claude-"))
-                            .map(|id| json!({ "name": id }))
-                            .collect();
-                        return json!(models);
-                    }
-                }
-            }
-            json!([])
-        }
-        "gemini" => {
-            let resp = client
-                .get(format!("https://generativelanguage.googleapis.com/v1/models?key={}", api_key))
-                .send()
-                .await;
-            if let Ok(r) = resp {
-                if let Ok(body) = r.json::<serde_json::Value>().await {
-                    if let Some(models) = body.get("models").and_then(|d| d.as_array()) {
-                        let filtered: Vec<Value> = models
-                            .iter()
-                            .filter_map(|m| m.get("name").and_then(|id| id.as_str()))
-                            .filter(|id| id.starts_with("models/"))
-                            .map(|id| json!({ "name": id.strip_prefix("models/").unwrap_or(id) }))
-                            .collect();
-                        return json!(filtered);
-                    }
-                }
-            }
-            json!([])
-        }
-        _ => json!([]),
-    }
-}
-
-async fn stream_openai_chat(
-    app: &AppHandle,
-    base_url: &str,
-    api_key: &str,
-    job_id: &str,
-    req: &AiGenerateRequest,
-) -> Result<(), String> {
-    let messages = req.messages.iter().map(|m| json!({ "role": m.role, "content": m.content })).collect::<Vec<_>>();
-    let temperature = req.temperature.unwrap_or(0.7);
-
-    let mut body = json!({
-        "model": req.model,
-        "messages": messages,
-        "stream": true,
-        "temperature": temperature,
-    });
-    if let Some(mt) = req.max_tokens {
-        body["max_tokens"] = json!(mt);
-    }
-
-    let mut response = reqwest::Client::builder()
-        .timeout(std::time::Duration::from_secs(300))
-        .build()
-        .map_err(|e| e.to_string())?
-        .post(format!("{base_url}/chat/completions"))
-        .bearer_auth(api_key)
-        .json(&body)
-        .send()
-        .await
-        .map_err(|e| format!("OpenAI unreachable: {e}"))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body_text = response.text().await.unwrap_or_default();
-        return Err(format!("OpenAI {status}: {body_text}"));
-    }
-
-    let mut line_buf = String::new();
-
-    loop {
-        // Check if job was cancelled before reading next chunk
-        if let Some(job) = app.state::<Mutex<JobTracker>>().lock().get(job_id) {
-            if job.status == crate::jobs::JobStatus::Cancelled {
-                drop(response);
-                return Err("Job cancelled".to_string());
-            }
-        }
-
-        match response.chunk().await {
-            Ok(Some(bytes)) => {
-                line_buf.push_str(&String::from_utf8_lossy(&bytes));
-
-                while let Some(nl) = line_buf.find('\n') {
-                    let line = line_buf[..nl].trim().to_string();
-                    line_buf = line_buf[nl + 1..].to_string();
-
-                    let data = match line.strip_prefix("data: ") {
-                        Some(d) => d.trim(),
-                        None => continue,
-                    };
-
-                    if data == "[DONE]" {
-                        let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": "", "done": true }));
-                        app.state::<Mutex<JobTracker>>().lock().complete(job_id, json!({ "done": true }));
-                        return Ok(());
-                    }
-
-                    let event: Value = match serde_json::from_str(data) {
-                        Ok(v) => v,
-                        Err(_) => continue,
-                    };
-
-                    let delta = event
-                        .get("choices")
-                        .and_then(|c| c.get(0))
-                        .and_then(|c| c.get("delta"))
-                        .and_then(|d| d.get("content"))
-                        .and_then(|c| c.as_str())
-                        .unwrap_or("");
-
-                    if !delta.is_empty() {
-                        let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": delta, "done": false }));
-                    }
-                }
-            }
-            Ok(None) => break,
-            Err(e) => return Err(format!("Stream error: {e}")),
-        }
-    }
-
-    let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": "", "done": true }));
-    app.state::<Mutex<JobTracker>>().lock().complete(job_id, json!({ "done": true }));
-    Ok(())
-}
-
-async fn stream_anthropic_chat(
-    app: &AppHandle,
-    api_key: &str,
-    job_id: &str,
-    req: &AiGenerateRequest,
-) -> Result<(), String> {
-    let temperature = req.temperature.unwrap_or(0.7);
-    let max_tokens = req.max_tokens.unwrap_or(4096);
-
-    let system_content: String = req.messages.iter()
-        .filter(|m| m.role == "system")
-        .map(|m| m.content.as_str())
-        .collect::<Vec<_>>()
-        .join("\n");
-    let messages: Vec<Value> = req.messages.iter()
-        .filter(|m| m.role != "system")
-        .map(|m| json!({ "role": m.role, "content": m.content }))
-        .collect();
-
-    // Enable extended thinking for balanced effort and above (max_tokens >= 2048).
-    // Thinking requires temperature=1; budget is half the output quota.
-    let thinking_budget = if max_tokens >= 2048 { max_tokens / 2 } else { 0 };
-    let actual_max_tokens = max_tokens + thinking_budget;
-
-    let mut body = json!({
-        "model": req.model,
-        "messages": messages,
-        "max_tokens": actual_max_tokens,
-        "stream": true,
-        "temperature": if thinking_budget > 0 { 1.0 } else { temperature },
-    });
-    if thinking_budget > 0 {
-        body["thinking"] = json!({ "type": "enabled", "budget_tokens": thinking_budget });
-    }
-    if !system_content.is_empty() {
-        body["system"] = json!(system_content);
-    }
-
-    let mut response = reqwest::Client::builder()
-        .timeout(std::time::Duration::from_secs(300))
-        .build()
-        .map_err(|e| e.to_string())?
-        .post("https://api.anthropic.com/v1/messages")
-        .header("x-api-key", api_key)
-        .header("anthropic-version", "2023-06-01")
-        .json(&body)
-        .send()
-        .await
-        .map_err(|e| format!("Anthropic unreachable: {e}"))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body_text = response.text().await.unwrap_or_default();
-        return Err(format!("Anthropic {status}: {body_text}"));
-    }
-
-    let mut line_buf = String::new();
-    let mut last_event = String::new();
-
-    loop {
-        // Check if job was cancelled before reading next chunk
-        if let Some(job) = app.state::<Mutex<JobTracker>>().lock().get(job_id) {
-            if job.status == crate::jobs::JobStatus::Cancelled {
-                drop(response);
-                return Err("Job cancelled".to_string());
-            }
-        }
-
-        match response.chunk().await {
-            Ok(Some(bytes)) => {
-                line_buf.push_str(&String::from_utf8_lossy(&bytes));
-
-                while let Some(nl) = line_buf.find('\n') {
-                    let line = line_buf[..nl].trim().to_string();
-                    line_buf = line_buf[nl + 1..].to_string();
-
-                    if let Some(event) = line.strip_prefix("event: ") {
-                        last_event = event.trim().to_string();
-                        continue;
-                    }
-
-                    let data = match line.strip_prefix("data: ") {
-                        Some(d) => d.trim(),
-                        None => continue,
-                    };
-
-                    if last_event == "message_stop" || data.contains("\"type\":\"message_stop\"") {
-                        let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": "", "done": true }));
-                        app.state::<Mutex<JobTracker>>().lock().complete(job_id, json!({ "done": true }));
-                        return Ok(());
-                    }
-
-                    let event: Value = match serde_json::from_str(data) {
-                        Ok(v) => v,
-                        Err(_) => continue,
-                    };
-
-                    let delta_obj = event.get("delta");
-                    let delta_type = delta_obj
-                        .and_then(|d| d.get("type"))
-                        .and_then(|t| t.as_str())
-                        .unwrap_or("");
-
-                    match delta_type {
-                        "thinking_delta" => {
-                            let thinking = delta_obj
-                                .and_then(|d| d.get("thinking"))
-                                .and_then(|t| t.as_str())
-                                .unwrap_or("");
-                            if !thinking.is_empty() {
-                                let _ = app.emit("ai:stream", json!({
-                                    "jobId": job_id,
-                                    "delta": thinking,
-                                    "done": false,
-                                    "thinking": true,
-                                }));
-                            }
-                        }
-                        "text_delta" => {
-                            let text = delta_obj
-                                .and_then(|d| d.get("text"))
-                                .and_then(|t| t.as_str())
-                                .unwrap_or("");
-                            if !text.is_empty() {
-                                let _ = app.emit("ai:stream", json!({
-                                    "jobId": job_id,
-                                    "delta": text,
-                                    "done": false,
-                                }));
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            Ok(None) => break,
-            Err(e) => return Err(format!("Stream error: {e}")),
-        }
-    }
-
-    let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": "", "done": true }));
-    app.state::<Mutex<JobTracker>>().lock().complete(job_id, json!({ "done": true }));
-    Ok(())
-}
-
-async fn stream_gemini_chat(
-    app: &AppHandle,
-    api_key: &str,
-    job_id: &str,
-    req: &AiGenerateRequest,
-) -> Result<(), String> {
-    let temperature = req.temperature.unwrap_or(0.7);
-
-    let system_text: String = req.messages.iter()
-        .filter(|m| m.role == "system")
-        .map(|m| m.content.as_str())
-        .collect::<Vec<_>>()
-        .join("\n");
-
-    let contents: Vec<Value> = req.messages.iter()
-        .filter(|m| m.role != "system")
-        .map(|m| {
-            let role = if m.role == "assistant" { "model" } else { "user" };
-            json!({ "role": role, "parts": [{ "text": m.content }] })
-        })
-        .collect();
-
-    let mut generation_config = json!({ "temperature": temperature });
-    if let Some(mt) = req.max_tokens {
-        generation_config["maxOutputTokens"] = json!(mt);
-    }
-    let mut body = json!({
-        "contents": contents,
-        "generationConfig": generation_config,
-    });
-    if !system_text.is_empty() {
-        body["systemInstruction"] = json!({ "parts": [{ "text": system_text }] });
-    }
-
-    let url = format!(
-        "https://generativelanguage.googleapis.com/v1beta/models/{}:streamGenerateContent?key={}",
-        req.model, api_key
-    );
-
-    let mut response = reqwest::Client::builder()
-        .timeout(std::time::Duration::from_secs(300))
-        .build()
-        .map_err(|e| e.to_string())?
-        .post(&url)
-        .json(&body)
-        .send()
-        .await
-        .map_err(|e| format!("Gemini unreachable: {e}"))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body_text = response.text().await.unwrap_or_default();
-        return Err(format!("Gemini {status}: {body_text}"));
-    }
-
-    let mut buf = String::new();
-    let mut depth: i32 = 0;
-    let mut in_string = false;
-    let mut escape = false;
-
-    loop {
-        // Check if job was cancelled before reading next chunk
-        if let Some(job) = app.state::<Mutex<JobTracker>>().lock().get(job_id) {
-            if job.status == crate::jobs::JobStatus::Cancelled {
-                drop(response);
-                return Err("Job cancelled".to_string());
-            }
-        }
-
-        match response.chunk().await {
-            Ok(Some(bytes)) => {
-                let chunk = String::from_utf8_lossy(&bytes).to_string();
-                for ch in chunk.chars() {
-                    if escape { escape = false; buf.push(ch); continue; }
-                    if ch == '\\' && in_string { escape = true; buf.push(ch); continue; }
-                    if ch == '"' { in_string = !in_string; }
-                    if !in_string {
-                        if ch == '{' { depth += 1; }
-                        else if ch == '}' { depth -= 1; }
-                    }
-                    buf.push(ch);
-
-                    if depth == 0 && buf.trim_start().starts_with('{') && !buf.trim().is_empty() {
-                        if let Ok(event) = serde_json::from_str::<Value>(buf.trim()) {
-                            let delta = event
-                                .get("candidates")
-                                .and_then(|c| c.get(0))
-                                .and_then(|c| c.get("content"))
-                                .and_then(|c| c.get("parts"))
-                                .and_then(|p| p.get(0))
-                                .and_then(|p| p.get("text"))
-                                .and_then(|t| t.as_str())
-                                .unwrap_or("");
-                            if !delta.is_empty() {
-                                let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": delta, "done": false }));
-                            }
-                        }
-                        buf.clear();
-                    }
-                }
-            }
-            Ok(None) => break,
-            Err(e) => return Err(format!("Stream error: {e}")),
-        }
-    }
-
-    let _ = app.emit("ai:stream", json!({ "jobId": job_id, "delta": "", "done": true }));
-    app.state::<Mutex<JobTracker>>().lock().complete(job_id, json!({ "done": true }));
-    Ok(())
+    json!(provider_for(&provider).list_models(&client, &api_key).await)
 }
 
 #[tauri::command]
